+++ src/main/java/peddle/controllers/WishListController.java
@@ -1,13 +1,17 @@
 package peddle.controllers;
 
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.ResponseBody;
 import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+
 
 import peddle.dto.EventDtoRs;
+import peddle.dto.WishListDto;
 import peddle.services.WishListService;
 
 import java.util.List;
@@ -21,12 +25,21 @@ public class WishListController {
 
   @GetMapping("/wishlist/user-events/{userId}")
   @ResponseBody
-  public List<EventDtoRs> getAllUserEvents(@PathVariable Long userId){
+  public List<EventDtoRs> getAllUserEvents(@PathVariable Long userId) {
     List<EventDtoRs> events = wishListService.getAllByUserId(userId);
     return events;
   }
 
+  @PostMapping("/wishlist/add-event-to-user")
+  @ResponseBody
+  public void addEventToUser(@RequestBody WishListDto wishListDto) {
+    wishListService.addNewEventToUser(wishListDto);
+  }
 
-
+  @PostMapping("/wishlist/delete-event-from-user")
+  @ResponseBody
+  public void deleteEventFromUser(@RequestBody WishListDto wishListDto) {
+    wishListService.deleteBadEventFromUser(wishListDto);
+  }
 
 }
