+++ src/main/java/peddle/repository/WishListRepository.java
@@ -3,6 +3,6 @@ package peddle.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import peddle.entities.WishList;
 
-public interface WishListRepository extends JpaRepository<WishList, Long>{
+public interface WishListRepository extends JpaRepository<WishList, Long> {
 
 }
