+++ src/main/java/peddle/services/EventServiceImpl.java
@@ -6,18 +6,13 @@ import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Service;
 import peddle.configuration.BadRequestException;
-import peddle.configuration.ErrorConstants;
-import peddle.configuration.UserException;
 
 import peddle.dto.EventDtoRs;
 import peddle.dto.EventDtoRq;
-import peddle.dto.UserEventDto;
 import peddle.dto.EventFullDtoRs;
 
 import peddle.entities.Event;
-import peddle.entities.User;
 import peddle.repository.EventRepository;
-import peddle.repository.UserRepository;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -36,12 +31,6 @@ public class EventServiceImpl implements EventService {
   @Autowired
   private ModelMapper modelMapper;
 
-  private final UserRepository userRepository;
-
-  public EventServiceImpl(UserRepository userRepository) {
-    this.userRepository = userRepository;
-  }
-
   @Override
   public List<EventDtoRs> getAll() {
     List<EventDtoRs> eventsDto = new ArrayList<>();
@@ -53,7 +42,7 @@ public class EventServiceImpl implements EventService {
   public List<EventDtoRs> getAllByPage(int page, int size) {
     List<EventDtoRs> eventsDto = new ArrayList<>();
     eventRepository.findAll(PageRequest.of(page, size, Sort.by(SORT_ORDER)))
-        .forEach(event -> eventsDto.add(modelMapper.map(event, EventDtoRs.class)));
+            .forEach(event -> eventsDto.add(modelMapper.map(event, EventDtoRs.class)));
     return eventsDto;
   }
 
@@ -61,67 +50,32 @@ public class EventServiceImpl implements EventService {
   public List<EventDtoRs> getByFilter(EventDtoRq eventDtoRq) {
     List<Event> events;
     PageRequest pageRequest = PageRequest.of(
-        eventDtoRq.getPage(),
-        eventDtoRq.getPageSize(),
-        Sort.by(SORT_ORDER));
+            eventDtoRq.getPage(),
+            eventDtoRq.getPageSize(),
+            Sort.by(SORT_ORDER));
 
     try {
       if (eventDtoRq.getCityName().isEmpty()) {
         events = eventRepository.findEventByDateBetween(
-            eventDtoRq.getDateStartConverted(),
-            eventDtoRq.getDateFinConverted(),
-            pageRequest);
+                eventDtoRq.getDateStartConverted(),
+                eventDtoRq.getDateFinConverted(),
+                pageRequest);
       } else {
         events = eventRepository.findEventByCity_NameAndDateBetween(
-            eventDtoRq.getCityName(),
-            eventDtoRq.getDateStartConverted(),
-            eventDtoRq.getDateFinConverted(),
-            pageRequest);
+                eventDtoRq.getCityName(),
+                eventDtoRq.getDateStartConverted(),
+                eventDtoRq.getDateFinConverted(),
+                pageRequest);
       }
     } catch (ParseException e) {
       throw new BadRequestException("Bad date in request");
     }
 
     List<EventDtoRs> eventsDtoRs = events.stream()
-        .map(event -> modelMapper.map(event, EventDtoRs.class))
-        .collect(Collectors.toList());
+            .map(event -> modelMapper.map(event, EventDtoRs.class))
+            .collect(Collectors.toList());
 
     return eventsDtoRs;
-  }
-
-  @Override
-    public List<EventDtoRs> getAllByUserId(Long userId) {
-    User user = userRepository.findById(userId)
-            .orElseThrow(() -> new UserException(ErrorConstants.ERR_USER_NOT_FOUND));
-    return user.getEvents().stream()
-            .map(event -> modelMapper.map(event, EventDtoRs.class)).collect(Collectors.toList());
-  }
-
-  @Override
-  public void addNewEventToUser(UserEventDto userEventDto) {
-    User user = userRepository.findById(userEventDto.getUserId())
-            .orElseThrow(() -> new UserException(ErrorConstants.ERR_USER_NOT_FOUND));
-
-    Event event = eventRepository.findById(userEventDto.getEventId())
-            .orElseThrow(() -> new UserException(ErrorConstants.ERR_EVENT_NOT_FOUND));
-
-    user.getEvents().add(event);
-    userRepository.save(user);
-
-    event.getUsers().add(user);
-    eventRepository.save(event);
-  }
-
-  @Override
-  public void deleteBadEventFromUser(UserEventDto userEventDto) {
-    User user = userRepository.findById(userEventDto.getUserId())
-            .orElseThrow(()-> new UserException(ErrorConstants.ERR_USER_NOT_FOUND));
-
-    Event event = eventRepository.findById(userEventDto.getEventId())
-            .orElseThrow(()-> new UserException(ErrorConstants.ERR_EVENT_NOT_FOUND));
-
-    user.getEvents().remove(event);
-    userRepository.save(user);
   }
 
   public EventFullDtoRs getById(Long id) {
